Add files via upload
Description: This project demonstrates the implementation of a basic linear regression model using TensorFlow and Keras. The model is trained to fit the function 𝑦=2𝑥−1
y=2x−1, where input-output pairs are used to train a neural network. The neural network consists of a single dense layer with one unit, and the training is done using mean squared error loss and stochastic gradient descent optimizer. Additionally, the project includes model evaluation and predictions on new data.